fix: update dashboard layout
